[#524] PDFInfoMenu 프로퍼티 정리
