+++ Presentation/MainPDF/PDFInfoMenu.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import PDFKit
 import UIKit
 
 //MARK: - ActivityView
@@ -32,7 +31,6 @@ public struct ActivityViewController: UIViewControllerRepresentable {
 //MARK: - PDFInfoMenu
 struct PDFInfoMenu: View {
     private let pdfSharedData: PDFSharedData = .shared
-    @State private var isActivityViewPresented = false
     @EnvironmentObject private var homeViewModel: HomeViewModel
     @EnvironmentObject private var mainPDFViewModel: MainPDFViewModel
     @EnvironmentObject private var pdfInfoMenuViewModel: PDFInfoMenuViewModel
@@ -41,17 +39,12 @@ struct PDFInfoMenu: View {
     @State var title: String?
     @State var isStarSelected: Bool = false
     
-    private let fileURL: URL = {
-        let fileName = PDFSharedData.shared.paperInfo?.title ?? "Untitled"
-        return FileManager.default.temporaryDirectory.appending(path: "\(fileName).pdf")
-    }()
 
     var body: some View {
         VStack(spacing: 12) {
             
             Button(action: {
-                setTemporaryPDF()
-                isActivityViewPresented = true
+                pdfInfoMenuViewModel.activityButtonTapped()
             }, label: {
                 HStack {
                     VStack(alignment: .leading) {
@@ -85,8 +78,8 @@ struct PDFInfoMenu: View {
                         .foregroundStyle(.gray100)
                 )
             })
-            .popover(isPresented: $isActivityViewPresented) {
-                ActivityViewController(activityItems: [self.fileURL])
+            .popover(isPresented: $pdfInfoMenuViewModel.isActivityViewPresented) {
+                ActivityViewController(activityItems: [pdfInfoMenuViewModel.fileURL])
             }
             
             VStack(spacing: 10) {
@@ -200,35 +193,6 @@ struct PDFInfoMenu: View {
         }
         .onDisappear {
             homeViewModel.changedTitle = nil
-        }
-    }
-    
-    
-    private func setTemporaryPDF() {
-        guard let document = pdfSharedData.document?.copy() as? PDFDocument else { return }
-        
-        for pageIndex in 0 ..< document.pageCount {
-            guard let page = document.page(at: pageIndex) else { continue }
-            
-            // 각 페이지의 모든 주석을 반복하며 밑줄과 코멘트 아이콘 지우기
-            for annotation in page.annotations {
-                guard let contents = annotation.contents else { continue }
-                
-                // 하이라이트의 contents가 "UH|"로 시작하면 지우지 않기
-                if !contents.hasPrefix("UH|") {
-                    page.removeAnnotation(annotation)
-                }
-            }
-        }
-        
-        // PDF 파일을 지정한 URL에 덮어쓰기 저장
-        do {
-            let pdfData = document.dataRepresentation()
-            try pdfData?.write(to: self.fileURL)
-            
-            print("PDF 저장이 완료되었습니다.")
-        } catch {
-            print("PDF 저장 중 오류 발생: \(error.localizedDescription)")
         }
     }
 }
