+++ Presentation/MainPDF/ViewModel/PDFInfoMenuViewModel.swift
@@ -6,10 +6,18 @@
 //
 
 import Foundation
+import PDFKit
 
 @MainActor
 class PDFInfoMenuViewModel: ObservableObject {
     private let pdfInfoMenuUsecase: PDFInfoMenuUseCase
+    
+    @Published public var isActivityViewPresented: Bool = false
+    
+    public var fileURL: URL {
+        let fileName = PDFSharedData.shared.paperInfo?.title ?? "Untitled"
+        return FileManager.default.temporaryDirectory.appending(path: "\(fileName).pdf")
+    }
     
     init(pdfInfoMenuUsecase: PDFInfoMenuUseCase) {
         self.pdfInfoMenuUsecase = pdfInfoMenuUsecase
@@ -36,4 +44,37 @@ class PDFInfoMenuViewModel: ObservableObject {
         }
     }
     
+    public func activityButtonTapped() {
+        setTemporaryPDF()
+        self.isActivityViewPresented.toggle()
+    }
+    
+    
+    private func setTemporaryPDF() {
+        guard let document = PDFSharedData.shared.document?.copy() as? PDFDocument else { return }
+        
+        for pageIndex in 0 ..< document.pageCount {
+            guard let page = document.page(at: pageIndex) else { continue }
+            
+            // 각 페이지의 모든 주석을 반복하며 밑줄과 코멘트 아이콘 지우기
+            for annotation in page.annotations {
+                guard let contents = annotation.contents else { continue }
+                
+                // 하이라이트의 contents가 "UH|"로 시작하면 지우지 않기
+                if !contents.hasPrefix("UH|") {
+                    page.removeAnnotation(annotation)
+                }
+            }
+        }
+        
+        // PDF 파일을 지정한 URL에 덮어쓰기 저장
+        do {
+            let pdfData = document.dataRepresentation()
+            try pdfData?.write(to: self.fileURL)
+            
+            print("PDF 저장이 완료되었습니다.")
+        } catch {
+            print("PDF 저장 중 오류 발생: \(error.localizedDescription)")
+        }
+    }
 }
